force cast to 64, huh

+++ lit_model.py
@@ -47,7 +47,7 @@ class LitModel(L.LightningModule):
             "current_hard_prompt",
             tokenizer.encode(
                 "Please multiply these two 3 digit numbers as best you possibly can. No talk; just go.",
-            ),
+            ).to(torch.int64),
         )
 
         assert (
